i can see a ghost hand on the hamburger on teh floor when i hover over it.

+++ aspirant_application/Graphics.AvatarInventory.cpp
@@ -8,6 +8,7 @@
 #include "Graphics.Fonts.h"
 #include "Game.Item.h"
 #include "Graphics.Data.Types.h"
+#include <optional>
 namespace graphics::Layouts
 {
 	nlohmann::json& GetLayout(const std::string&);
@@ -140,7 +141,7 @@ namespace graphics::AvatarInventory
 		}, []() {});
 	}
 
-	std::optional<int> OnMouseButtonUp(const std::string& layoutName, const std::string& controlId, const common::XY<Sint32>& xy, Uint8 buttons)
+	std::optional<size_t> OnMouseButtonUp(const std::string& layoutName, const std::string& controlId, const common::XY<Sint32>& xy, Uint8 buttons)
 	{
 		return WithControl<std::optional<int>>(layoutName, controlId,
 			[xy, buttons](nlohmann::json& thingie) 
@@ -155,10 +156,10 @@ namespace graphics::AvatarInventory
 				auto inventory = game::avatar::Items::All();
 				if ((size_t)row < inventory.size())
 				{
-					return std::optional<int>(row);
+					return std::optional<size_t>(row);
 				}
 			}
-			return std::optional<int>();
+			return std::optional<size_t>();
 		}, []() { return std::nullopt; });
 	}
 }
